metavariable-pattern: Fix nested language implementation
1. Pass the right xlang to `mini_rule_of_pattern`, this fixes a crash
(`Impossible` exception) when using a non-generic pattern within a
generic rule.

2. Adjust parse info after `with_tmp_file`, this fixes the ranges given
of the matches obtained by metavariable-pattern/language (which dumps
the content of the metavariable into a temporary file in order to call
the parsing functions), so that these refer to the correct original file
and position. This alllows to start a patterns op with a pattern-not.

Fixes: 70a20605ae4 ("metavariable-pattern: Allow matching content under another language (#3366)")

test plan:
make test # tests included